plugins/obs: Rewrite against websocket v5 and libdex
This is, effectively, and almost from-scratch rewrite of ObsConnection.
It works against obs-websocket v5 and is entirely based on libdex.

The reason for using libdex here is because websocket v5 forces us
to do many more synchronous requests, and with the previous code
each request was split into a caller and a callback, so the chain
of execution was already difficult to follow. With v5 it became
a complete mess. With libdex, it's incredibly easier to follow, at
the cost of more code setting things up and more lifetime management.

Part-of: <https://gitlab.gnome.org/World/boatswain/-/merge_requests/56>