Introduce region-aware pages
Make BsPage handle regions, and have items per region. Add a versioning
field to BsPage so we can migrate more easily in the future.

The new page layout looks like this:

```
{
  "version" : 1,
  "regions" : [
    {
      "id" : "main-button-grid",
      "region-data" : {            },
      "items" : [
        ...
      ]
    },
    {
      "id" : "touchscreen",
      "region-data" : {            },
      "items" : [
        ...
      ]
    }
  ]
}
```

Because this is a big change and I'm not 100% sure the code is correct,
add some code to create backups of the config files.

Related: https://gitlab.gnome.org/World/boatswain/-/issues/80