Introduce BsDeviceUpdate
It's a simple transactional mechanism that queues updates in a list -
one list for buttons, one list for touchscreens - and allows them to
be queried later.

The update object has two states: it initialized as a mutable object
and collects updated; then it has to be sealed, i.e. made immutable,
before the updates can be queried.