elgato/device-provider: Port to DexChannel
Add 2 channels, one for added devices, and one for removed devices.

Channels are thread-safe ways to transfer data. In this case, the
data is GUsbDevice objects. We send these objects on GUsbContext
callbacks - which are executed in separate threads - and the receiver
end is a fiber running in the main thread.